+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unpublished
 ### Added
-- add period backer overview export (CSV only)
+- add period backer overview
+- add period backer assignment with quick backer creation component
+### Fixed
+- error of empty backer list enabled elements
 
 ## v0.7.13 - 2024-05-28
 ### Added